in stock-exchange-service adding getCompaniesListForAStockExchange method and creating new microservice sector-service

+++ backend/stock-exchange-service/src/main/java/com/ud/stockexchangeservice/repositories/StockExchangeRepository.java
@@ -3,5 +3,5 @@ package com.ud.stockexchangeservice.repositories;
 import com.ud.stockexchangeservice.entities.StockExchange;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StockExchangeRepository extends JpaRepository<StockExchange, Integer> {
+public interface StockExchangeRepository extends JpaRepository<StockExchange, Long> {
 }
